src/api/requirements.txt
flask
redis
pytest
psycopg2-binary
dependency-injector
d6tstack
geomet
geopy
tqdm
joblib
psutil
pyproj
shapely
simplejson
pqdm
pillow